Refactor: separate service into interface and implementation
